requirements.txt
numpy
pip>=9.0.0
pandas
matplotlib
Keras==2.1.6
h5py
pillow
scikit-learn
scipy
tensorboard==1.8.0
tensorflow==1.8.0
